atlas_api/requirements.txt
google-auth
google-cloud-bigquery
pandas